revert: retry policy of unconfirmed transaction from stream (#326)
This reverts commit d59cd1d21494d62a61997d5aec13f60f61d87c12.
